requirements.txt
flask
pillow
pytesseract
transformers
torch
